+++ R/data.R
@@ -4,7 +4,26 @@
 #' supported by this package.
 #' 
 #' @format A numeric vector with 2 entries
+#' 
+#' @examples
+#' valid_iea_release_years
 "valid_iea_release_years"
+
+
+#' IEA data frame columns
+#'
+#' A string list containing named names of columns in IEA data frames.
+#' These data frames can be 
+#' tidy (one row per data point) or
+#' wide (years spread to the right).
+#' Names of the items in the list are the typical function arguments
+#' for column names.
+#' 
+#' @format A string list with 10 entries
+#' 
+#' @examples
+#' iea_cols
+"iea_cols"
 
 
 #' Coal and coal products
@@ -12,6 +31,9 @@
 #' A string vector containing names of products classified by the IEA as coal and coal products.
 #' 
 #' @format A string vector with 16 entries
+#' 
+#' @examples
+#' coal_and_coal_products
 "coal_and_coal_products"
 
 
@@ -20,6 +42,8 @@
 #' A string vector containing names of products classified by the IEA as primary coal products.
 #' 
 #' @format A string vector with 7 entries
+#' @examples
+#' primary_coal_products
 "primary_coal_products"
 
 
@@ -28,6 +52,9 @@
 #' A string vector containing names of products classified by the IEA as `coal_and_coal_products` that are not `primary_coal_products`.
 #' 
 #' @format A string vector with 9 entries
+#' 
+#' @examples 
+#' secondary_coal_products
 "secondary_coal_products"
 
 
@@ -36,6 +63,9 @@
 #' A string vector containing names of products classified by the IEA as peat and peat products.
 #' 
 #' @format A string vector with 2 entries
+#' 
+#' @examples 
+#' peat_and_peat_products
 "peat_and_peat_products"
 
 
@@ -44,6 +74,9 @@
 #' A string vector containing names of products classified by the IEA as primary peat products.
 #' 
 #' @format A string vector with 1 entry
+#' 
+#' @examples 
+#' primary_peat_products
 "primary_peat_products"
 
 
@@ -52,6 +85,9 @@
 #' A string vector containing names of products classified by the IEA as "Peat and peat products" that are not `primary_peat_products`.
 #' 
 #' @format A string vector with 1 entry
+#' 
+#' @examples 
+#' secondary_peat_products
 "secondary_peat_products"
 
 
@@ -60,6 +96,9 @@
 #' A string vector containing names of products classified by the IEA as oil and oil products.
 #' 
 #' @format A string vector with 23 entries
+#' 
+#' @examples 
+#' oil_and_oil_products
 "oil_and_oil_products"
 
 
@@ -68,6 +107,9 @@
 #' A string vector containing names of products classified by the IEA as primary oil products.
 #' 
 #' @format A string vector with 2 entries
+#' 
+#' @examples 
+#' primary_oil_products
 "primary_oil_products"
 
 
@@ -76,6 +118,9 @@
 #' A string vector containing names of products classified by the IEA as `oil_and_oil_products` that are not `primary_coal_products`.
 #' 
 #' @format A string vector with 18 entries
+#' 
+#' @examples 
+#' secondary_oil_products
 "secondary_oil_products"
 
 
@@ -84,6 +129,9 @@
 #' A string vector containing names of products classified by the IEA as renewables.
 #' 
 #' @format A string vector with 23 entries
+#' 
+#' @examples 
+#' renewable_products
 "renewable_products"
 
 
@@ -92,6 +140,9 @@
 #' A string vector containing names of products classified by the IEA as biofuel and waste products.
 #' 
 #' @format A string vector with 10 entries
+#' 
+#' @examples
+#' biofuels_and_waste_products
 "biofuels_and_waste_products"
 
 
@@ -100,6 +151,9 @@
 #' A string vector containing names of `Total primary energy supply` `Flow`s in the IEA extended energy balances database.
 #' 
 #' @format A string vector with 6 entries
+#' 
+#' @examples 
+#' tpes_flows
 "tpes_flows"
 
 
@@ -108,6 +162,9 @@
 #' A string vector containing names of `Total final consumption` comparison `Flow`s in the IEA extended energy balances database.
 #' 
 #' @format A string vector with 6 entries
+#' 
+#' @examples 
+#' tfc_compare_flows
 "tfc_compare_flows"
 
 
@@ -116,6 +173,9 @@
 #' A string vector containing names of `Total final consumption` `Flow`s in the IEA extended energy balances database.
 #' 
 #' @format A string vector with 4 entries
+#' 
+#' @examples 
+#' tfc_flows
 "tfc_flows"
 
 
@@ -124,6 +184,9 @@
 #' A string vector containing names of `Industry` `Flow`s in the IEA extended energy balances database.
 #' 
 #' @format A string vector with 15 entries
+#' 
+#' @examples
+#' industry_flows
 "industry_flows"
 
 
@@ -132,6 +195,9 @@
 #' A string vector containing names of `Manufacturing` `Flow`s in the IEA extended energy balances database.
 #' 
 #' @format A string vector with 15 entries
+#' 
+#' @examples
+#' manufacturing_flows
 "manufacturing_flows"
 
 
@@ -140,6 +206,9 @@
 #' A string vector containing names of `Transport` `Flow`s in the IEA extended energy balances database.
 #' 
 #' @format A string vector with 8 entries
+#' 
+#' @examples
+#' transport_flows
 "transport_flows"
 
 
@@ -148,6 +217,9 @@
 #' A string vector containing names of `Other` `Flow`s in the IEA extended energy balances database.
 #' 
 #' @format A string vector with 6 entries
+#' 
+#' @examples
+#' other_flows
 "other_flows"
 
 
@@ -156,6 +228,9 @@
 #' A string vector containing names of industries whose purpose in IEA extended energy balance `Flow`s is to provide aggregations.
 #' 
 #' @format A string vector with 8 entries
+#' 
+#' @examples
+#' aggregation_flows
 "aggregation_flows"
 
 #' Memo and aggregation flow prefixes
@@ -163,6 +238,9 @@
 #' A string vector containing names of `Flow`s that provide memos and aggregations.
 #' 
 #' @format A string vector with 3 entries
+#' 
+#' @examples
+#' memo_aggregation_flow_prefixes
 "memo_aggregation_flow_prefixes"
 
 #' Memo and aggregation product prefixes
@@ -170,6 +248,9 @@
 #' A string vector containing names of `Product`s that provide memos and aggregations.
 #' 
 #' @format A string vector with 2 entries
+#' 
+#' @examples
+#' memo_aggregation_product_prefixes
 "memo_aggregation_product_prefixes"
 
 #' Interface industries
@@ -177,7 +258,54 @@
 #' A string vector containing names of `Flow`s that interface with the world outside of the economy.
 #' 
 #' @format A string vector with 5 entries
+#' 
+#' @examples
+#' interface_industries
 "interface_industries"
+
+
+#' Country order
+#'
+#' A string vector containing 3-letter ISO country codes in alphabetical order.
+#' 
+#' @format A string vector
+#' 
+#' @examples
+#' country_order
+"country_order"
+
+
+#' Method order
+#'
+#' A string vector containing types of methods for quantifying the primary equivalent of renewable electricity.
+#' 
+#' @format A string vector
+#' 
+#' @examples
+#' method_order
+"method_order"
+
+
+#' Energy type order
+#'
+#' A string vector containing the order for energy types.
+#' 
+#' @format A string vector
+#' 
+#' @examples
+#' energy_type_order
+"energy_type_order"
+
+
+#' Last stage order
+#'
+#' A string vector containing options for the last stage of energy conversion chain analysis.
+#' 
+#' @format A string vector
+#' 
+#' @examples
+#' last_stage_order
+"last_stage_order"
 
 
 #' IEA order for ledger side
@@ -185,6 +313,9 @@
 #' A string vector containing `Ledger.side` entries in IEA order.
 #' 
 #' @format A string vector
+#' 
+#' @examples
+#' ledger_side_iea_order
 "ledger_side_iea_order"
 
 
@@ -193,6 +324,9 @@
 #' A string vector containing entries for a united flow aggregation point and flow column in IEA order.
 #' 
 #' @format A string vector
+#' 
+#' @examples
+#' fap_flow_iea_order
 "fap_flow_iea_order"
 
 
@@ -201,4 +335,7 @@
 #' A string vector containing entries for the product column in IEA order.
 #' 
 #' @format A string vector
+#' 
+#' @examples
+#' product_iea_order
 "product_iea_order"